Moving classes into more appropriate packages
