+++ src/main/java/requests/FloorRequest.java
@@ -1,4 +1,4 @@
-package misc;
+package requests;
 
 import systemwide.Direction;
