+++ src/main/java/requests/ServiceRequest.java
@@ -1,4 +1,4 @@
-package misc;
+package requests;
 
 import systemwide.Direction;
