+++ src/test/java/scheduler/SchedulerTest.java
@@ -2,9 +2,8 @@ package scheduler;
 
 import elevatorsystem.ElevatorSubsystem;
 import floorsystem.FloorSubsystem;
-import misc.BoundedBuffer;
-import misc.ElevatorRequest;
-import misc.ServiceRequest;
+import systemwide.BoundedBuffer;
+import requests.ServiceRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
