+++ src/test/java/systemwide/BoundedBufferTest.java
@@ -1,10 +1,11 @@
-package misc;
+package systemwide;
 
 import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
-import systemwide.Direction;
+
+import requests.ServiceRequest;
 
 /**
  * Test class for BoundedBuffer methods
