+++ src/test/java/systemwide/DirectionTest.java
@@ -1,3 +1,5 @@
+package systemwide;
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import systemwide.Direction;
